// backend/src/resources/enums.ts
export enum HttpMessages {
    DATA_ADDED_SUCCESSFULLY = "Data added successfully",
    DATA_DELETED_SUCCESSFULLY = "Data deleted successfully",
    DATA_EDITED_SUCCESSFULLY = "Data edited successfully",
    DATA_RETRIEVED_SUCCESSFULLY = "Data retrieved successfully",
    NO_DATA_FOUND = "No data found",
    ERROR_ADDING_DATA = "Error adding data",
    ERROR_DELETING_DATA = "Error deleting data",
    ERROR_EDITING_DATA = "Error editing data",
    ERROR_GETTING_DATA = "Error getting data",
    INVALID_DATA_FORMAT = "Invalid data format",
    MISSING_REQUIRED_FIELDS = "Missing required fields"
}